Cast away signed conversion warnings

+++ ReactiveCocoaFramework/ReactiveCocoa/fishhook/fishhook.c
@@ -64,7 +64,7 @@ static void perform_rebinding_with_section(struct section *section,
                                            char *strtab,
                                            uint32_t *indirect_symtab) {
   uint32_t *indirect_symbol_indices = indirect_symtab + section->reserved1;
-  void **indirect_symbol_bindings = (void **)(slide + section->addr);
+  void **indirect_symbol_bindings = (void **)((uintptr_t)slide + section->addr);
   for (int i = 0; i < section->size / sizeof(void *); i++) {
     uint32_t symtab_index = indirect_symbol_indices[i];
     int32_t strtab_offset = symtab[symtab_index].n_un.n_strx;
@@ -90,7 +90,7 @@ static void rebind_symbols_for_image(const struct mach_header *header,
   if (dladdr(header, &info) == 0) {
     return;
   }
-  intptr_t cur = (intptr_t)header + sizeof(struct mach_header);
+  uintptr_t cur = (uintptr_t)header + sizeof(struct mach_header);
   struct segment_command* cur_seg_cmd;
   struct segment_command* linkedit_segment = NULL;
   struct section* lazy_symbols = NULL;
@@ -128,7 +128,7 @@ static void rebind_symbols_for_image(const struct mach_header *header,
     return;
   }
   // Find base symbol/string table addresses
-  uint32_t linkedit_base = slide + linkedit_segment->vmaddr - linkedit_segment->fileoff;
+  uint32_t linkedit_base = (uintptr_t)slide + linkedit_segment->vmaddr - linkedit_segment->fileoff;
   struct nlist* symtab = (struct nlist *)(linkedit_base + symtab_cmd->symoff);
   char *strtab = (char *)(linkedit_base + symtab_cmd->stroff);
   // Get indirect symbol table (array of uint32_t indices into symbol table)
